Add direct game launching.

+++ BootNTR/source/clock.c
@@ -17,9 +17,9 @@ u64     Timer_Restart(void)
 u64		getTimeInMsec(u64 timer) {
 	return timer / CPU_TICKS_PER_MSEC;
 }
-bool    Timer_HasTimePassed(u32 nbmsecToWait, u64 timer)
+bool    Timer_HasTimePassed(float nbmsecToWait, u64 timer)
 {
-	u64 seconds = nbmsecToWait * CPU_TICKS_PER_MSEC;
+	u64 seconds = (u64)(nbmsecToWait * CPU_TICKS_PER_MSEC);
 	u64 current = svcGetSystemTick();
 	u64 diff = current - timer;
