+++ BootNTR/source/clock.h
@@ -2,5 +2,5 @@
 #include <3ds.h>
 
 u64     Timer_Restart(void);
-bool    Timer_HasTimePassed(u32 nbmsecToWait, u64 timer);
+bool    Timer_HasTimePassed(float nbmsecToWait, u64 timer);
 u64		getTimeInMsec(u64 timer);
